Move `useSubscriberEmailAddress` to `@automattic/data-stores/src/reader/hooks`
