+++ packages/data-stores/src/reader/hooks/index.ts
@@ -1,5 +1,7 @@
 import { useSelect } from '@wordpress/data';
+import { useMemo } from '@wordpress/element';
 import { register as registerUserStore } from '../../user';
+import { getSubkey } from '../helpers';
 import type { UserSelect } from '../../user';
 
 const USER_STORE = registerUserStore( { client_id: '', client_secret: '' } );
@@ -18,4 +20,25 @@ export const useIsQueryEnabled = () => {
 	}
 
 	return false;
+};
+
+// Get subscriber's email address based on the subkey cookie
+export const useSubscriberEmailAddress = () => {
+	return useMemo( () => {
+		const subkey = getSubkey();
+
+		if ( ! subkey ) {
+			return null;
+		}
+
+		const decodedSubkeyValue = decodeURIComponent( subkey );
+
+		const firstPeriodIndex = decodedSubkeyValue.indexOf( '.' );
+		if ( firstPeriodIndex === -1 ) {
+			return null;
+		}
+
+		const emailAddress = decodedSubkeyValue.slice( firstPeriodIndex + 1 );
+		return emailAddress;
+	}, [] );
 };
